requirements.test.txt
pytest
pytest-cov
pytest-homeassistant-custom-component
aiohttp>=3.8.0
pytest-asyncio
pre-commit